recoloring, camera control, finetuning

+++ Util/M.cs
@@ -29,7 +29,7 @@ namespace Wyri
 
         public static double Euclidean(Point p1, Point p2) => Math.Sqrt(Math.Pow(p1.X - p2.X, 2) + Math.Pow(p1.Y - p2.Y, 2));
 
-        public static double Euclidean(Vector2 p1, Vector2 p2) => Math.Sqrt(Math.Pow(p1.X - p2.X, 2) + Math.Pow(p1.Y - p2.Y, 2));
+        public static float Euclidean(Vector2 p1, Vector2 p2) => (float)Math.Sqrt(Math.Pow(p1.X - p2.X, 2) + Math.Pow(p1.Y - p2.Y, 2));
 
         public static double Euclidean(float p1x, float p1y, float p2x, float p2y) => Math.Sqrt(Math.Pow(p1x - p2x, 2) + Math.Pow(p1y - p2y, 2));
